Clean the username on mounts directory. (#2845)
* Clean the username on mounts directory.

* Fix mypy and pylint check failed.
